# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List the signed-in user's uploaded documents
      responses:
        "200":
          description: The user's documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
    post:
      operationId: registerDocument
      tags: [documents]
      summary: Register an uploaded document and extract its text
      description: >
        Called after a file is uploaded to object storage. The server downloads
        it, extracts readable text (txt, pdf, docx), and stores the document so
        the assistant can reference it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDocumentInput"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMeta"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete one of the signed-in user's documents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /documents/extract:
    post:
      operationId: extractDocumentText
      tags: [documents]
      summary: Read the text out of an uploaded document without saving it
      description: >
        Called after a TXT, Markdown, Word, or PDF file is uploaded to object
        storage. The server downloads it and extracts its readable text,
        returning that text so it can be inserted inline into a reflection,
        the Mind profiler, the assistant chat, or any other text field — for
        example attaching a report the user wrote as context. Does not persist
        the document to the user's library.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractDocumentTextInput"
      responses:
        "200":
          description: Text read from the document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractDocumentTextResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
  /ocr/extract:
    post:
      operationId: extractImageText
      tags: [ocr]
      summary: Read the text out of an uploaded image (OCR)
      description: >
        Called after a screenshot or photo is uploaded to object storage. The
        server downloads it and uses a vision model to transcribe any text it
        contains, returning that text so it can be inserted into a to-do, the
        Mind profiler, the assistant chat, or any other text field as proof or
        context. If the image holds no legible text, returns an empty string.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractImageTextInput"
      responses:
        "200":
          description: Text read from the image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractImageTextResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
        "500":
          description: OCR failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
components:
  schemas:
...
    ApiErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DocumentMeta:
      type: object
      description: Metadata about an uploaded document.
      properties:
        id:
          type: string
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
        charCount:
          type: integer
          description: Number of characters of text extracted from the document.
        createdAt:
          type: string
      required:
        - id
        - name
        - contentType
        - size
        - charCount
        - createdAt
    DocumentList:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentMeta"
      required:
        - documents
    RegisterDocumentInput:
      type: object
      properties:
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
        objectPath:
          type: string
          description: The /objects/... path returned by the upload step.
      required:
        - name
        - contentType
        - objectPath
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    ExtractImageTextInput:
      type: object
      properties:
        objectPath:
          type: string
          description: The /objects/... path of the uploaded image.
      required:
        - objectPath
    ExtractDocumentTextInput:
      type: object
      properties:
        objectPath:
          type: string
          description: The /objects/... path of the uploaded document.
        name:
          type: string
          description: The original filename, used to pick the right text parser.
        contentType:
          type: string
          description: The uploaded file's MIME type.
      required:
        - objectPath
        - name
        - contentType
    ExtractDocumentTextResult:
      type: object
      properties:
        text:
          type: string
          description: The text read from the document. Empty if none was found.
      required:
        - text
    ExtractImageTextResult:
      type: object
      properties:
        text:
          type: string
          description: The text read from the image. Empty if none was legible.
      required:
        - text
...