# lib/api-zod/src/generated/types/documentList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentMeta } from './documentMeta';

export interface DocumentList {
  documents: DocumentMeta[];
}