# lib/api-zod/src/generated/types/extractDocumentTextInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractDocumentTextInput {
  /** The /objects/... path of the uploaded document. */
  objectPath: string;
  /** The original filename, used to pick the right text parser. */
  name: string;
  /** The uploaded file's MIME type. */
  contentType: string;
}