# lib/api-zod/src/generated/types/extractDocumentTextResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractDocumentTextResult {
  /** The text read from the document. Empty if none was found. */
  text: string;
}