# lib/api-zod/src/generated/types/extractImageTextInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractImageTextInput {
  /** The /objects/... path of the uploaded image. */
  objectPath: string;
}